+++ src/App.js
@@ -1,10 +1,12 @@
 import "./App.css";
-import Home from "./Components/Pages/Home.tsx";
+import Home from "./Components/pages/Home";
 
 function App() {
-  return <div className="App">
-    <Home></Home>
-  </div>;
+  return (
+    <div className="App">
+      <Home />
+    </div>
+  );
 }
 
 export default App;
